requirements.txt
anthropic>=0.34.0
# Core dependencies
anthropic>=0.18.0
pdfplumber>=0.10.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# CLI and formatting
click>=8.1.0
tabulate>=0.9.0

# Data processing
pandas>=2.0.0
regex>=2023.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
python-dotenv>=1.0.0
